Ramakrishna_Ashram_chatbot/requirement.txt
fastapi
uvicorn
langgraph
langchain
qdrant-client
python-dotenv
openai
langchain-openai
langchain-qdrant
PyMuPDF
